TripX/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
httpx>=0.27.0
pydantic>=2.7.4
google-genai>=0.2.0
